// api/github-sls-rest-api/api.ts
export interface GithubSlsRestApiChallenge {
    /**
     * 
     * @type {string}
     * @memberof GithubSlsRestApiChallenge
     */
    'recipient'?: string;
    /**
     * 
     * @type {string}
     * @memberof GithubSlsRestApiChallenge
     */
    'invitation'?: string;
    /**
     * 
     * @type {Array<GithubSlsRestApiTotpMethod>}
     * @memberof GithubSlsRestApiChallenge
     */
    'methods': Array<GithubSlsRestApiTotpMethod>;
    /**
     * 
     * @type {string}
     * @memberof GithubSlsRestApiChallenge
     */
    'org': string;
}
// ...
export const GithubSlsRestApiTotpMethod = {
    App: 'app',
    Email: 'email'
} as const;

export type GithubSlsRestApiTotpMethod = typeof GithubSlsRestApiTotpMethod[keyof typeof GithubSlsRestApiTotpMethod];